Warn when extending built-ins.
Most people agree it's not a great idea to
extend built-in objects. With this option
we will warn when built-in objects are
extended in the following ways:

    Object.prototype = {};
	Object.prototype.property = function () {};

There are several other ways that devs can
extend natives that are not covered, but this
seems like it covers the majority of cases.

This currently relies on the list in constants.js
labeled `ecmaIdentifiers` which is imperfect but
probably works well enough.